move container orchestrator acceptance tests to their own class
